1inch token resolver - refactored getTokenData
